+++ HISTORY.rst
@@ -5,10 +5,11 @@ History
 
 3.7.0 (2023-??-??)
 ------------------
-This version introduces new filtering functions, a new optional parameter for Principal Component Analysis, improvements to the graphical interface, and bug fixes.
+This version introduces small RNA read alignment using ShortStack, new filtering functions, a new optional parameter for Principal Component Analysis, improvements to the graphical interface, and bug fixes.
 
 Added
 *******
+* Added small RNA read alignment using ShortStack (fastq.shortstack_align_smallrna).
 * Added new filtering function 'Filter specific rows by name' (Filter.filter_by_row_name).
 * Added new filtering function 'Filter rows with duplicate names/IDs' (Filter.filter_duplicate_ids).
 * Added new parameter to Principal Component Analysis (CountFilter.pca) 'proportional_axes', that allows you to make the PCA projection axes proportional to the percentage of variance each PC explains.